% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/estimate.R
\name{bayes.rule}
\alias{bayes.rule}
\title{Bayes rule for estimating posteriors.}
\usage{
bayes.rule(X, densities, pi, type = c("sum", "prod"))
}
\arguments{
\item{X}{Matrix of predictors.}

\item{type}{Calculate posteriors using raw products of probabilities or summation of logs of
probabilities? "Sum" may offer superior floating point accuracy.}
}
\description{
Bayes rule for estimating posteriors.
}

