% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calculate.discriminants}
\alias{calculate.discriminants}
\title{Calculate discriminants}
\usage{
calculate.discriminants(X, parameters, type = c("linear", "quadratic",
  "poisson"))
}
\description{
Wrapper for discriminant functions. Not currently in use.
}

