% Generated by roxygen2 (4.0.2): do not edit by hand
\name{estimate.densities}
\alias{estimate.densities}
\title{Estimate densities.}
\usage{
estimate.densities(X, mu.hat, sigma.hat, method = "linear")
}
\arguments{
\item{X}{Matrix of predictors.}

\item{method}{Method of discriminant analysis to be
performed. Linear assumes that all population class covariances are
equal. Quadratic estimates each population class covariance
separately.}
}
\description{
Estimate densities.
}

