% Generated by roxygen2 (4.0.2): do not edit by hand
\name{estimate.normal.densities}
\alias{estimate.normal.densities}
\title{Estimate normal densities.}
\usage{
estimate.normal.densities(X, mu.hat, sigma.hat, type = c("linear",
  "quadratic"))
}
\arguments{
\item{X}{Matrix of predictors.}

\item{type}{Type of discriminant analysis to be performed. Linear assumes that all population
class covariances are equal. Quadratic estimates each population class covariance
separately. Linear and Quadratic assume predictors are normally distributed (i.e. continous
data). Poisson models count data.}
}
\description{
Estimate normal densities.
}

