% Generated by roxygen2 (4.0.2): do not edit by hand
\name{estimate.poisson.densities}
\alias{estimate.poisson.densities}
\title{Estimate poisson densities.}
\usage{
estimate.poisson.densities(X, N.hat, d.hat)
}
\arguments{
\item{X}{Matrix of predictors.}
}
\description{
Estimate poisson densities.
}

