% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate.poisson.seq.densities}
\alias{estimate.poisson.seq.densities}
\title{Estimate poisson sequence densities.}
\usage{
estimate.poisson.seq.densities(X, N.hat, d.hat)
}
\arguments{
\item{X}{Matrix of predictors.}
}
\description{
Estimate poisson sequence densities.
}

