% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate.size.factor}
\alias{estimate.size.factor}
\title{Estimate size factor.}
\usage{
estimate.size.factor(X, X.train.parameter = NULL, type = c("mle",
  "quantile", "medratio"))
}
\arguments{
\item{X}{Matrix of predictors.}

\item{X.train.parameter}{Statistic of the training data used in calculating new s.hats. Rather
than calculate this from scratch using the training data, just remember it for later. This
parameter is different for each type.}

\item{type}{Type of discriminant analysis to be performed. Linear assumes that all population
class covariances are equal. Quadratic estimates each population class covariance
separately. Linear and Quadratic assume predictors are normally distributed (i.e. continous
data). Poisson models count data.}
}
\description{
Estimate size factor.
}

