% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/estimate.R
\name{fit.posteriors}
\alias{fit.posteriors}
\title{Fit posteriors.}
\usage{
fit.posteriors(posteriors, levels)
}
\arguments{
\item{posteriors}{N by K matrix of posterior probabilities.}

\item{levels}{K-length vector of factor levels corresponding to response classes.}
}
\value{
N-length vector of predicted classes.
}
\description{
Fit posteriors.
}

