% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_synth_corpus.R
\name{gen_synth_corpus}
\alias{gen_synth_corpus}
\title{Generates a Corpus with a given \eqn{(\alpha, \eta)}}
\usage{
gen_synth_corpus(K, V, num.docs, doc.size, alpha.h, eta.h)
}
\arguments{
\item{K}{number of topics}

\item{V}{vocabulary size}

\item{doc.size}{number of words in each document}

\item{J}{number of collections}

\item{collection.size}{number of documents in each collection (a list of J
elements)}

\item{alpha.vec}{a set of values for document-level Dirichlet sampling}

\item{eta.vec}{a set of values for topic Diriclet sampling}
}
\value{
a list of generated corpus and their statistics
}
\description{
Generates documents using the LDA generative process based on a set of
\eqn{(\alpha, \eta)} Configurations
}
\details{
Last modified on: April 15, 2016
}
\examples{
## Generates documents with given parameters

K                  <- 4
V                  <- 20
alpha.h            <- .2
eta.h              <- .5
doc.size           <- 80
num.docs           <- 40 # number of documents in each collection
ds.name            <- paste("synth-K", K, "-V", V, sep = "")

ds                 <- gen_synth_corpus(K, V, num.docs, doc.size, alpha.h, eta.h)

}
\seealso{
Other corpus: 
\code{\link{calc_doc_cos}()},
\code{\link{calc_doc_tf}()},
\code{\link{gen_synth_corpus_given_beta}()}
}
\concept{corpus}
