% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_synth_corpus.R
\name{gen_synth_corpus_given_beta}
\alias{gen_synth_corpus_given_beta}
\title{Generates a Corpus with a given \eqn{(\alpha, \beta)}}
\usage{
gen_synth_corpus_given_beta(K, V, num.docs, doc.size, alpha.h, beta.samples)
}
\arguments{
\item{K}{number of topics}

\item{V}{vocabulary size}

\item{doc.size}{number of words in each document}

\item{beta.samples}{the topic matrix beta}

\item{J}{number of collections}

\item{collection.size}{number of documents in each collection (a list of J
elements)}

\item{alpha.vec}{a set of values for document-level Dirichlet sampling}
}
\value{
a list of generated corpus and their statistics
}
\description{
Generates documents using the LDA generative process based on a set of
\eqn{(\alpha, \beta)} Configurations
}
\details{
Last modified on: July 12, 2016
}
\seealso{
Other corpus: 
\code{\link{calc_doc_cos}()},
\code{\link{calc_doc_tf}()},
\code{\link{gen_synth_corpus}()}
}
\concept{corpus}
