% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_rearrange_column.R
\name{greedyRearrangeColumn}
\alias{greedyRearrangeColumn}
\title{Rearrange columns of two matrices for match in a greedy way}
\usage{
greedyRearrangeColumn(mat1, mat2)
}
\arguments{
\item{mat1}{Matrix to align to.}

\item{mat2}{Matrix to be rearranged.}
}
\value{
Vector of order to be applied to columns of mat2.
}
\description{
Rearrange columns of two matrices for match in a greedy way
}
\examples{
\dontrun{
column_order <- greedyRearrangeColumn(mat1, mat2)
mat2 <- mat2[, column_order]
}
}
