% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{hkGenes}
\alias{hkGenes}
\title{Human housekeeping genes.}
\format{A data frame with 4061 rows and 2 variables:
\describe{
  \item{ensembl_gene_id}{Ensemble gene id.}
  \item{hgnc_symbol}{HGNC sympol.}
}}
\source{
\url{https://doi.org/10.1016/j.tig.2013.05.010}
\url{https://www.tau.ac.il/~elieis/HKG/}
}
\usage{
hkGenes
}
\description{
A dataset containing 4061 human housekeeping genes.
}
\keyword{datasets}
