% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrange_column.R
\name{PermutationRearrangeColumn}
\alias{PermutationRearrangeColumn}
\title{Rearrange columns in mat2 to align with mat1, using l2 distance}
\usage{
PermutationRearrangeColumn(mat1, mat2)
}
\arguments{
\item{mat1}{Matrix to align to.}

\item{mat2}{Matrix to be rearranged.}
}
\value{
Vector of order to be applied to columns of mat2.
}
\description{
Rearrange columns in mat2 to align with mat1, using l2 distance
}
\examples{
\dontrun{
column_order <- PermutationRearrangeColumn(mat1, mat2)
mat2 <- mat2[, column_order]
}
}
