% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plda}
\alias{plda}
\title{Poisson discriminant analysis.}
\usage{
plda(X, y, method = c("linear", "quadratic"), dist = c("normal", "poisson"))
}
\arguments{
\item{X}{Matrix of predictors.}

\item{y}{Factor vector of class or group labels.}

\item{method}{Method of discriminant analysis to be
performed. Linear assumes that all population class covariances are
equal. Quadratic estimates each population class covariance
separately.}

\item{dist}{Underlying distribution of predictors. Normal models
continuous data. Poisson models count data.}
}
\value{
List of class "plda" containing the classifier results.
}
\description{
Poisson discriminant analysis.
}

