% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plda_choose_lambda.R
\name{plda_choose_lambda}
\alias{plda_choose_lambda}
\title{pLDA Choose Shrinkage Parameter}
\usage{
plda_choose_lambda(dat, k, lam_values, idx_hk, njobs = 1, n_rep = 1,
  fdir = NA, seed = 2019)
}
\arguments{
\item{dat}{Input matrix. A matrix of positive integers where rows represent genes (words) and columns represent cells (documents).}

\item{k}{Number of topics.}

\item{lam_values}{A vector of shrinkage parameter values.}

\item{idx_hk}{Index (column number) of housekeeping gene.}

\item{njobs}{Number of cores used for the calculation. Default 1.}

\item{n_rep}{Number of repetition for each lam_values. Default 1.}

\item{fdir}{Output file directory. Default ../output/}

\item{seed}{LDA seed. Default 2019. If a vector, use the best seed in terms of likelihood.}
}
\value{
Plot number of interesting genes for different lambda values. Return table of lam_values and corresponding number of interesting genes.
}
\description{
Run pLDA under different lambda values. Plot and return output. Interesting genes are defined as those with variance in beta matrix greater than average variance of housekeeping genes.
}
\examples{
\dontrun{
plda_choose_lambda(dat=cell_by_gene_expr_matrix, k=10, lam_values=c(10, 10^2, 10^3),  idx_hk = 1:10)
}
}
