% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plda_estimate_beta.R
\name{plda_estimate_beta_given_gamma}
\alias{plda_estimate_beta_given_gamma}
\title{Estimate beta matrix given data and gamma matrix}
\usage{
plda_estimate_beta_given_gamma(x, theta.true)
}
\arguments{
\item{x}{Input matrix. A matrix of positive integers where rows represents cells (documents) and column represents genes (words). Here the input matrix should be associated with gamma.}

\item{gamma}{Topic by sample matrix gamma associated with input matrix x.}
}
\value{
log_beta matrix based on x and gamma.
}
\description{
Calculate topic by gene (beta) matrix given data and gamma matrix.
}
\examples{
\dontrun{
plda_estimate_beta(log_beta=plda_model$logProbW, x=new_expression_data)
}
}
