% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plda_estimate_gamma.R
\name{plda_estimate_gamma_given_beta}
\alias{plda_estimate_gamma_given_beta}
\title{Estimate new gamma matrix}
\usage{
plda_estimate_gamma_given_beta(x, log_beta, seed = 2018,
  verbose = FALSE)
}
\arguments{
\item{x}{Input matrix. A matrix of positive integers where rows represents cells (documents) and column represents genes (words).}

\item{log_beta}{Estimated log beta matrix from previous run of pLDA. Number of topics by vocabulary size.}
}
\value{
Gamma matrix for the new expression data.
}
\description{
Calculate topic by cell (gamma) matrix for new expression data
with estimated topic by gene matrix.
}
\examples{
\dontrun{
plda_estimate_gamma_given_beta(x=new_expression_data, log_beta=plda_model$logProbW)
}
}
