% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gamma.R
\name{plotGamma}
\alias{plotGamma}
\title{Plot heatmap for gamma matrix}
\usage{
plotGamma(fit)
}
\arguments{
\item{fit}{Model fit from plda()}
}
\value{
Heatmap of gamma matrix.
}
\description{
Plot heatmap for gamma matrix
}
\examples{
\dontrun{
plotGamma(fit=plda_model)
}
}
