% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distributions.R
\name{poisson.seq.discriminant}
\alias{poisson.seq.discriminant}
\title{Poisson sequence discriminant.}
\usage{
poisson.seq.discriminant(x, s, g, d, pi)
}
\description{
Poisson sequence discriminant.
}

