% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess pLDA input data}
\usage{
preprocess(data, var.mean.cutoff = 1, row.max.cutoff = 10)
}
\arguments{
\item{data}{Input matrix. A matrix of positive integers where rows represents cells (documents) and column represents genes (words).}

\item{var.mean.cutoff}{Row variance/mean cutoff.}

\item{row.max.cutoff}{Row max cutoff.}
}
\value{
Preprocessed data.
}
\description{
Preprocess pLDA input data. Keep only genes with across samples variance/mean > var.mean.cutoff and across samples max > row.max.cutoff.
}
\examples{
\dontrun{
preprocess(data=cell_by_gene_expr_matrix)
}
}
