% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{simulate.poisson}
\alias{simulate.poisson}
\title{Simulate Poisson.}
\usage{
\method{simulate}{poisson}(n, p, K, lambda)
}
\value{
List of X and y.
X is (n*K by p) data matrix composed of independent univariate poissons.
y is (n by 1) factor of class labels.
}
\description{
Simulate Poisson.
}

