% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_counts.R
\name{simulateCounts}
\alias{simulateCounts}
\title{Simulate expression count table}
\usage{
simulateCounts(theta, beta, doc_length)
}
\arguments{
\item{theta}{Theta matrix where rows represent cells and columns represent topics.}

\item{beta}{Beta matrix where rows represent topics and columns represent genes.}

\item{doc_length}{Library size of cell samples.}
}
\value{
A cell by gene matrix of expression counts.
}
\description{
Simulate expression count table given topic by gene, topic by cell matrices and library size.
}
\examples{
\dontrun{
counts <- simulateCounts(theta.true, beta.true, rep(100000, nrow(theta.true)))
}
}
