% Generated by roxygen2 (4.0.2): do not edit by hand
\name{split.data}
\alias{split.data}
\title{Split data.}
\usage{
\method{split}{data}(x, train = 0.7, val = 0, test = 0.3)
}
\arguments{
\item{x}{Matrix or vector to split.}

\item{train}{Proportion of rows to label as train.}

\item{val}{Proportion of rows to label as validation.}

\item{test}{Proportion of rows to label as test.}
}
\value{
List of split indicies.
}
\description{
Split data.
}

