% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate.poisson.densities}
\alias{estimate.poisson.densities}
\title{Estimate poisson densities.}
\usage{
estimate.poisson.densities(X, lambda.hat)
}
\arguments{
\item{X}{Matrix of predictors.}
}
\description{
Estimate poisson densities.
}

