% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plda}
\alias{plda}
\title{Poisson discriminant analysis.}
\usage{
plda(X, y, type = c("linear", "quadratic", "poisson"),
  size.factor = c("mle", "quantile", "medratio"), prior = c("uniform",
  "proportion"))
}
\arguments{
\item{X}{Matrix of predictors.}

\item{y}{Factor vector of class or group labels.}

\item{type}{Type of discriminant analysis to be performed. Linear assumes that all population
class covariances are equal. Quadratic estimates each population class covariance
separately. Linear and Quadratic assume predictors are normally distributed (i.e. continous
data). Poisson models count data.}
}
\value{
List of class "plda" containing the classifier results.
}
\description{
Poisson discriminant analysis.
}

