% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plda_functions.R
\name{plda}
\alias{plda}
\title{Penalized Latent Dirichlet Allocation}
\usage{
plda(x, k, lambda, warm.start.seed = 2017, verbose = FALSE,
  warm.start.lda = NULL, warm.start.log.beta = NULL, pdata = NULL)
}
\arguments{
\item{x}{Input matrix. A matrix of positive integers where rows represents cells (documents) and column represents genes (words).}

\item{k}{Number of topics.}

\item{lambda}{Shrinkage parameter.}

\item{warm.start.seed}{Seed of LDA estimation used as pLDA warm start. pLDA uses parameters estimated from LDA as a warm start to boost speed.}

\item{verbose}{Default FALSE. Print progress.}

\item{warm.start.lda}{Default NULL. LDA object for warm start.}

\item{warm.start.log.beta}{Default NULL. log beta matrix for warm start.}

\item{pdata}{Default NULL. Table of phenotypic data.}
}
\value{
plda() returns a list of the penalized LDA output. logProbW is the log of topic by gene matrix beta. gamma is the cell by topic matrix of topic frequencies for each cell. genes are gene names extracted from the column name of input count matrix.
}
\description{
The main function that runs penalized LDA.
}
\examples{
\dontrun{
plda(x=cell_by_gene_expr_matrix, k=10, lambda=5)
}
}
